mom-ai/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
numpy>=1.24.0
torch>=2.0.0
grpcio>=1.65.0
protobuf>=5.0.0
